xbot/requirements.txt
httpx>=0.27
bpx-py @ git+https://github.com/sndmndss/bpx-py.git
lighter-sdk>=0.1.4
grvt-pysdk>=0.2.1
pytest>=8.2

pytest-asyncio>=0.23
